Add unit tests and CI
