+++ .github/workflows/test.yml
@@ -1,0 +1,16 @@
+name: Test
+
+on:
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-node@v3
+        with:
+          node-version: '18'
+      - run: npm ci --ignore-scripts --no-audit --no-fund || true
+      - run: npm test
