+++ index.html
@@ -114,6 +114,7 @@
         </div>
     </div>
 
+    <script src="utils.js"></script>
     <script src="script.js"></script>
 </body>
